Moved and renamed file output class
